backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
requests>=2.31.0
# ML/Data Science Dependencies
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
opencv-python>=4.8.0
h5py>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scipy>=1.11.0
httpx>=0.24.0
groq
